% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLogit.R
\name{CLogit}
\alias{CLogit}
\title{Implements the CLogit Family for matched
case-control data, to be used with \code{mboost::gamboost}.}
\usage{
CLogit(risktype = c("classic", "median"))
}
\arguments{
\item{risktype}{Character string. Either \code{"classic"} (mean risk)
or \code{"median"} (median risk).}
}
\value{
An object of class \code{Family} usable in \code{gamboost}.
}
\description{
Implements the CLogit Family for matched
case-control data, to be used with \code{mboost::gamboost}.
}
