% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_continuous}
\alias{detect_continuous}
\title{Helper: detect continuous variables}
\usage{
detect_continuous(df, exclude = c())
}
\arguments{
\item{df}{A data.frame containing all data.}
}
\value{
Two lists, continuous variables, and categorical ones.
}
\description{
Helper: detect continuous variables
}
