% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interaction_const}
\alias{interaction_const}
\title{Fit an offset boosting model for stratified/matched data
Checks if the two variables lead to a constant interaction.}
\usage{
interaction_const(x, z)
}
\arguments{
\item{x}{Character name of first column.}

\item{z}{Character name of second column.}
}
\value{
TRUE/FALSE, if the two variables lead to constant interaction.
}
\description{
Fit an offset boosting model for stratified/matched data
Checks if the two variables lead to a constant interaction.
}
