% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset.R
\name{gen_offset_model}
\alias{gen_offset_model}
\title{Fit an offset boosting model for stratified/matched data}
\usage{
gen_offset_model(
  data,
  formula,
  mstop,
  nu,
  strata,
  n_cores = 1,
  K = 5,
  early_stopping = TRUE,
  steady_state_percentage = 0.01,
  plot = TRUE,
  save_cv = FALSE
)
}
\arguments{
\item{data}{A data.frame containing the outcome, covariates, and strata variable.}

\item{formula}{A \code{gamboost} formula defining predictors for the offset model.}

\item{mstop}{Integer maximum number of boosting iterations.}

\item{nu}{Numeric step size for boosting (learning rate, default 1).}

\item{strata}{Character string naming the strata variable for matched design.}

\item{n_cores}{Integer number of how many cores are available for CV.}

\item{K}{Integer of folds for cross-validation (default 5).}

\item{early_stopping}{Logical. Toggle early stopping on or off.}

\item{steady_state_percentage}{Integer Threshold for minimal improvement in CV risk
to declare the model in steady state (default 0.01, i.e., 0.01\% change).}

\item{plot}{Logical. If the CV of the offset should be printed.}

\item{save_cv}{Logical. It indicates whether the cross-validated early stopping
object should be saved for analysis.}
}
\value{
A fitted \code{gamboost} object with optimal number of iterations.
}
\description{
Fits a component-wise gradient boosting model to estimate an offset
for the main effect model. Uses cross-validation to determine
the optimal number of boosting iterations (\code{mstop}).
}
\examples{
\dontrun{
offset_mod <- gen_offset_model(data, formula = "resp ~ X + Z1 + Z2",
mstop = 500, nu = 0.1, strata = "strata")
}
}
